assignment2

+++ csharp3/csharp3/Program.cs
@@ -16,7 +16,7 @@
             int sum = num1 + num2;
             int difference = num1 - num2;
             int multiplication = num1 * num2;
-            double division = (double)num1 / num2;
+            decimal division = (decimal)num1 / num2;
 
             Console.WriteLine("Suma: {0} + {1} = {2}", num1, num2, sum);
             Console.WriteLine("Skirtumas: {0} - {1} = {2}", num1, num2, difference);
